edited recepies page
