+++ src/features/RecepiesPage/Recepies.module.scss
@@ -4,53 +4,3 @@
 	background-color: $main-color;
 	padding: 250px calc(50% - 670px);
 }
-
-.header {
-	display: flex;
-	justify-content: space-evenly;
-	align-items: center;
-	width: 100%;
-	margin-bottom: 100px;
-}
-
-.bars {
-	position: relative;
-}
-
-.cta_btn {
-	position: absolute;
-	right: -5px;
-	top: 50%;
-	transform: translate(-50%, -50%);
-	cursor: pointer;
-	border: none;
-	background-color: transparent;
-}
-
-.category_icon {
-	font-size: 16px;
-}
-
-.bar {
-	box-shadow: 3px 3px 24px 0px rgba(76, 76, 76, 0.26);
-	width: 300px;
-	padding: 15px;
-	border-radius: 5px;
-	font-family: 'Montserrat', sans-serif;
-	appearance: none;
-}
-
-
-.icon {
-	font-size: 16px;
-}
-
-.recepies {
-	display: grid;
-	grid-template-columns: repeat(auto-fill, minmax(250px, 1fr));
-	row-gap: 100px;
-}
-
-.btn {
-	text-transform: uppercase;
-}
