+++ src/features/RecepiesPage/RecepiesHeader/RecepiesHeader.module.scss
@@ -1,0 +1,31 @@
+.header {
+	// display: grid;
+	// grid-template-columns: repeat(3, 1fr);
+	// justify-items: center;
+    display: flex;
+    gap: 60px;
+	width: 100%;
+	margin-bottom: 100px;
+}
+
+.searchbar {
+	position: relative;
+    width: 100%;
+}
+
+.searchbar_icon {
+	position: absolute;
+	right: 5px;
+	top: 50%;
+	transform: translate(-50%, -50%);
+	
+}
+
+.select {
+    width: 100%;
+}
+
+.btn {
+	text-transform: uppercase;
+    min-width: 180px;
+}
