+++ src/features/RecepiesPage/RecepiesList/RecepiesList.module.scss
@@ -1,0 +1,5 @@
+.recepies {
+	display: grid;
+	grid-template-columns: repeat(auto-fill, minmax(250px, 1fr));
+	row-gap: 100px;
+}
